requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0
django-debug-toolbar==4.3.0
django-filter==23.5
djangorestframework==3.14.0
docutils==0.20.1
drf-spectacular==0.27.1
flake8==7.0.0
flake8-docstrings==1.7.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
pillow==10.2.0
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
snowballstemmer==2.2.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
